+++ hi-ml/src/health_ml/utils/common_utils.py
@@ -22,17 +22,22 @@ MAX_PATH_LENGTH = 260
 empty_string_to_none = lambda x: None if (x is None or len(x.strip()) == 0) else x
 string_to_path = lambda x: None if (x is None or len(x.strip()) == 0) else Path(x)
 
-EXPERIMENT_SUMMARY_FILE = "experiment_summary.txt"
-CHECKPOINT_FOLDER = "checkpoints"
+# file and directory names
 CHECKPOINT_SUFFIX = ".ckpt"
 AUTOSAVE_CHECKPOINT_FILE_NAME = "autosave"
 AUTOSAVE_CHECKPOINT_CANDIDATES = [AUTOSAVE_CHECKPOINT_FILE_NAME + CHECKPOINT_SUFFIX,
                                   AUTOSAVE_CHECKPOINT_FILE_NAME + "-v1" + CHECKPOINT_SUFFIX]
-RUN_RECOVERY_ID_KEY = 'run_recovery_id'
-EFFECTIVE_RANDOM_SEED_KEY_NAME = "effective_random_seed"
-RUN_RECOVERY_FROM_ID_KEY_NAME = "recovered_from"
+CHECKPOINT_FOLDER = "checkpoints"
 DEFAULT_AML_UPLOAD_DIR = "outputs"
 DEFAULT_LOGS_DIR_NAME = "logs"
+EXPERIMENT_SUMMARY_FILE = "experiment_summary.txt"
+
+# run recovery
+RUN_RECOVERY_ID_KEY = 'run_recovery_id'
+RUN_RECOVERY_FROM_ID_KEY_NAME = "recovered_from"
+
+# other
+EFFECTIVE_RANDOM_SEED_KEY_NAME = "effective_random_seed"
 
 
 @unique
@@ -281,3 +286,13 @@ def set_model_to_eval_mode(model: Module) -> Generator:
     model.eval()
     yield
     model.train(old_mode)
+
+
+def is_long_path(path: PathOrString) -> bool:
+    """
+    A long path is a path that has more than MAX_PATH_LENGTH characters
+
+    :param path: The path to check the length of
+    :return: True if the length of the path is greater than MAX_PATH_LENGTH, else False
+    """
+    return len(str(path)) > MAX_PATH_LENGTH