+++ hi-ml/testhiml/testhiml/test_run_ml.py
@@ -1,40 +1,67 @@
+import shutil
 from pathlib import Path
 
 import pytest
-from typing import Tuple
-from unittest.mock import patch, MagicMock, Mock
+from typing import Generator, Tuple
+from unittest.mock import patch
 
-from pytorch_lightning import Callback
-
+from health_ml.configs.hello_container import HelloContainer
 from health_ml.experiment_config import ExperimentConfig
 from health_ml.lightning_container import LightningContainer
 from health_ml.run_ml import MLRunner
 
 
-@pytest.fixture
-def ml_runner() -> MLRunner:
-    experiment_config = ExperimentConfig()
+@pytest.fixture(scope="module")
+def ml_runner_no_setup() -> MLRunner:
+    experiment_config = ExperimentConfig(model="HelloContainer")
     container = LightningContainer(num_epochs=1)
-    return MLRunner(experiment_config=experiment_config, container=container)
+    runner = MLRunner(experiment_config=experiment_config, container=container)
+    return runner
 
 
-def test_ml_runner_setup(ml_runner: MLRunner) -> None:
-    """
-    Check that all the necessary methods get called during setup
-    """
-    assert not ml_runner._has_setup_run
-    with patch.object(ml_runner, "container", spec=LightningContainer) as mock_container:
+@pytest.fixture(scope="module")
+def ml_runner() -> Generator:
+    experiment_config = ExperimentConfig(model="HelloContainer")
+    container = LightningContainer(num_epochs=1)
+    runner = MLRunner(experiment_config=experiment_config, container=container)
+    runner.setup()
+    yield runner
+    output_dir = runner.container.file_system_config.outputs_folder
+    if output_dir.exists():
+        shutil.rmtree(output_dir)
+
+
+@pytest.fixture(scope="module")
+def ml_runner_with_container() -> Generator:
+    experiment_config = ExperimentConfig(model="HelloContainer")
+    container = HelloContainer()
+    runner = MLRunner(experiment_config=experiment_config, container=container)
+    runner.setup()
+    yield runner
+    output_dir = runner.container.file_system_config.outputs_folder
+    if output_dir.exists():
+        shutil.rmtree(output_dir)
+
+
+def _mock_model_train(chekpoint_path: Path, container: LightningContainer) -> Tuple[str, str]:
+    return "trainer", "storing_logger"
+
+
+def test_ml_runner_setup(ml_runner_no_setup: MLRunner) -> None:
+    """Check that all the necessary methods get called during setup"""
+    assert not ml_runner_no_setup._has_setup_run
+    with patch.object(ml_runner_no_setup, "container", spec=LightningContainer) as mock_container:
         with patch("health_ml.run_ml.seed_everything") as mock_seed:
-            # mock_container.get_effectie_random_seed = Mock()
-            ml_runner.setup()
+            ml_runner_no_setup.setup()
             mock_container.get_effective_random_seed.assert_called_once()
             mock_container.setup.assert_called_once()
             mock_container.create_lightning_module_and_store.assert_called_once()
-            assert ml_runner._has_setup_run
+            assert ml_runner_no_setup._has_setup_run
             mock_seed.assert_called_once()
 
 
 def test_set_run_tags_from_parent(ml_runner: MLRunner) -> None:
+    """Test that set_run_tags_from_parents causes set_tags to get called"""
     with pytest.raises(AssertionError) as ae:
         ml_runner.set_run_tags_from_parent()
         assert "should only be called in a Hyperdrive run" in str(ae)
@@ -47,47 +74,64 @@ def test_set_run_tags_from_parent(ml_runner: MLRunner) -> None:
 
 
 def test_run(ml_runner: MLRunner) -> None:
+    """Test that model runner gets called """
+    ml_runner.setup()
+    assert not ml_runner.checkpoint_handler.has_continued_training
+    with patch.object(ml_runner, "run_inference"):
+        with patch.object(ml_runner, "checkpoint_handler"):
+            with patch("health_ml.run_ml.model_train", new=_mock_model_train):
+                ml_runner.run()
+                assert ml_runner._has_setup_run
+                # expect _mock_model_train to be called and the result of ml_runner.storing_logger
+                # updated accordingly
+                assert ml_runner.storing_logger == "storing_logger"
+                assert ml_runner.checkpoint_handler.has_continued_training
 
-    def _mock_model_train(chekpoint_path: Path, container: LightningContainer) -> Tuple[str, str]:
-        return "trainer", dummy_storing_logger
 
-    dummy_storing_logger = "storing_logger"
-
-    with patch("health_ml.run_ml.model_train", new=_mock_model_train):
-        ml_runner.run()
-        assert ml_runner._has_setup_run
-        # expect _mock_model_train to be called and the result of ml_runner.storing_logger
-        # updated accordingly
-        assert ml_runner.storing_logger == dummy_storing_logger
-
-
-@patch("health_ml.run_ml.create_lightning_trainer")
-def test_run_inference_for_lightning_models(mock_create_trainer: MagicMock, ml_runner: MLRunner,
-                                            tmp_path: Path) -> None:
+def test_run_inference(ml_runner_with_container: MLRunner, tmp_path: Path) -> None:
     """
-    Check that all expected methods are called during inference3
+    Test that run_inference gets called as expected.
     """
-    mock_trainer = MagicMock()
-    mock_test_result = [{"result": 1.0}]
-    mock_trainer.test.return_value = mock_test_result
-    mock_create_trainer.return_value = mock_trainer, ""
+    def _expected_files_exist() -> int:
+        output_dir = ml_runner_with_container.container.outputs_folder
+        expected_files = [Path("test_mse.txt"), Path("test_mae.txt")]
+        return sum([p.exists() for p in expected_files] + [output_dir.is_dir()])
 
-    with patch.object(ml_runner, "container") as mock_container:
-        mock_container.num_gpus_per_node.return_value = 0
-        mock_container.get_trainer_arguments.return_value = {"callbacks": Callback()}
-        mock_container.load_model_checkpoint.return_value = Mock()
-        mock_container.get_data_module.return_value = Mock()
-        mock_container.pl_progress_bar_refresh_rate = None
-        mock_container.detect_anomaly = False
-        mock_container.pl_limit_train_batches = 1.0
-        mock_container.pl_limit_val_batches = 1.0
-        mock_container.outputs_folder = tmp_path
+    # create the test data
+    import numpy as np
+    import torch
 
-        checkpoint_paths = [Path("dummy")]
-        result = ml_runner.run_inference_for_lightning_models(checkpoint_paths)
-        assert result == mock_test_result
+    N = 100
+    x = torch.rand((N, 1)) * 10
+    y = 0.2 * x + 0.1 * torch.randn(x.size())
+    xy = torch.cat((x, y), dim=1)
+    data_path = tmp_path / "hellocontainer.csv"
+    np.savetxt(data_path, xy.numpy(), delimiter=",")
 
-        mock_create_trainer.assert_called_once()
-        mock_container.load_model_checkpoint.assert_called_once()
-        mock_container.get_data_module.assert_called_once()
-        mock_trainer.test.assert_called_once()
+    expected_ckpt_path = ml_runner_with_container.container.outputs_folder / "checkpoints" / "last.ckpt"
+    assert not expected_ckpt_path.exists()
+    # update the container to look for test data at this location
+    ml_runner_with_container.container.local_dataset_dir = tmp_path
+    assert _expected_files_exist() == 0
+
+    actual_train_ckpt_path = ml_runner_with_container.checkpoint_handler.get_recovery_or_checkpoint_path_train()
+    assert actual_train_ckpt_path is None
+    ml_runner_with_container.run()
+    actual_train_ckpt_path = ml_runner_with_container.checkpoint_handler.get_recovery_or_checkpoint_path_train()
+    assert actual_train_ckpt_path == expected_ckpt_path
+
+    actual_test_ckpt_path = ml_runner_with_container.checkpoint_handler.get_checkpoints_to_test()
+    assert actual_test_ckpt_path == [expected_ckpt_path]
+    assert actual_test_ckpt_path[0].exists()
+    # After training, the outputs directory should now exist
+    assert _expected_files_exist() == 3
+
+    # if no checkpoint handler, no checkpoint paths will be saved and these are required for
+    # inference so ValueError will be raised
+    with pytest.raises(ValueError) as e:
+        ml_runner_with_container.checkpoint_handler = None  # type: ignore
+        ml_runner_with_container.run()
+        assert "expects exactly 1 checkpoint for inference, but got 0" in str(e)
+
+    Path("test_mae.txt").unlink()
+    Path("test_mse.txt").unlink()